[Distributed] [CustomDevices] adapt lora sp && polish MC2 APIs
